+++ Sources/NablaMessagingUI/Scenes/Conversation/UI/Plugins/ConversationActivityPlugin/ConversationActivityCellProvider.swift
@@ -12,6 +12,7 @@ final class ConversationActivityCellProvider: ConversationCellProvider {
         collectionView: UICollectionView,
         indexPath: IndexPath,
         item: ConversationViewItem,
+        viewController _: UIViewController?,
         delegate: ConversationCellPresenterDelegate
     ) -> UICollectionViewCell? {
         guard let item = item as? ConversationActivityViewItem else {