+++ Sources/NablaMessagingUI/Scenes/Conversation/UI/Plugins/ConversationCellProvider.swift
@@ -9,6 +9,7 @@ protocol ConversationCellProvider {
         collectionView: UICollectionView,
         indexPath: IndexPath,
         item: ConversationViewItem,
+        viewController: UIViewController?,
         delegate: ConversationCellPresenterDelegate
     ) -> UICollectionViewCell?
 }