+++ Sources/NablaMessagingUI/Scenes/Conversation/UI/Plugins/DateSeparatorPlugin/DateSeparatorCellProvider.swift
@@ -12,6 +12,7 @@ final class DateSeparatorCellProvider: ConversationCellProvider {
         collectionView: UICollectionView,
         indexPath: IndexPath,
         item: ConversationViewItem,
+        viewController _: UIViewController?,
         delegate: ConversationCellPresenterDelegate
     ) -> UICollectionViewCell? {
         guard let item = item as? DateSeparatorViewItem else {