+++ Sources/NablaMessagingUI/Scenes/Conversation/UI/Plugins/HasMoreIndicatorPlugin/HasMoreIndicatorCellProvider.swift
@@ -12,6 +12,7 @@ final class HasMoreIndicatorCellProvider: ConversationCellProvider {
         collectionView: UICollectionView,
         indexPath: IndexPath,
         item: ConversationViewItem,
+        viewController _: UIViewController?,
         delegate: ConversationCellPresenterDelegate
     ) -> UICollectionViewCell? {
         guard let item = item as? HasMoreIndicatorViewItem else {