+++ Sources/NablaMessagingUI/Scenes/Conversation/UI/Plugins/MessagePlugin/Audio/AudioMessageCellProvider.swift
@@ -26,6 +26,7 @@ final class AudioMessageCellProvider: ConversationCellProvider {
         collectionView: UICollectionView,
         indexPath: IndexPath,
         item: ConversationViewItem,
+        viewController _: UIViewController?,
         delegate: ConversationCellPresenterDelegate
     ) -> UICollectionViewCell? {
         guard let item = item as? AudioMessageViewItem else {