+++ Sources/NablaMessagingUI/Scenes/Conversation/UI/Plugins/MessagePlugin/TypingIndicator/TypingIndicatorCellProvider.swift
@@ -14,6 +14,7 @@ final class TypingIndicatorCellProvider: ConversationCellProvider {
         collectionView: UICollectionView,
         indexPath: IndexPath,
         item: ConversationViewItem,
+        viewController _: UIViewController?,
         delegate: ConversationCellPresenterDelegate
     ) -> UICollectionViewCell? {
         guard let item = item as? TypingIndicatorViewItem else {