+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Added a "play" indicator on videos on iOS 16 since the default one has been removed by the system update.
+
 ### Fixed